fixup! Add HQC keymgmt
